requirements.txt
selenium==4.41.0
webdriver-manager==4.0.2
discord.py==2.7.1
discord-webhook==1.4.1
python-dotenv==1.0.0
